requirements.txt
asgiref
attrs
beautifulsoup4
bs4
certifi
charset-normalizer
Django
emoji
h11
idna
lxml
outcome
pillow
psycopg2-binary
PySocks
python-dateutil
requests
selenium
six
sniffio
sortedcontainers
soupsieve
sqlparse
trio
trio-websocket
typing_extensions
urllib3
wsproto
lxml
